refactor: complete config module extraction (Phase 2.3)
Extract all config functionality into organized submodules:

- config/loader.ts (153 lines): File I/O operations
  - loadGlobalConfig(), loadWorkspaceConfig(), loadLocalConfig()
  - saveWorkspaceConfig(), initConfig()
  - Path constants (CONFIG_FILES, GLOBAL_CONFIG_DIR, etc.)

- config/validator.ts (73 lines): Config validation
  - validateConfig() with provider, tool, pattern validation

- config/merger.ts (205 lines): Config merging with priority
  - mergeConfig() - merges global, workspace, local, CLI configs
  - DEFAULT_CONFIG - default configuration values
  - Helper functions to reduce code duplication

- config/utils.ts (147 lines): Utility functions
  - shouldAutoApprove(), isToolDisabled(), getToolDefaults()
  - mergeToolInput(), getExampleConfig(), getCustomDangerousPatterns()

- config/index.ts (61 lines): Module re-exports

Result:
- config.ts reduced from 855 to 54 lines (94% reduction)
- Total config module: 972 lines across 6 well-organized files
- Clean separation of concerns
- Backwards compatible imports
